--- modules/blog/blog-home/index.tsx
@@ -1,3 +1,0 @@
-import BlogHome from './blog-home';
-
-export default BlogHome;
