--- modules/blog/blog-posts-list/index.tsx
@@ -1,3 +1,0 @@
-import BlogPostsList from './blog-posts-list';
-
-export default BlogPostsList;
