+++ modules/blog/post-snippet/index.tsx
@@ -1,0 +1,3 @@
+import PostSnippet from './post-snippet';
+
+export default PostSnippet;
