+++ modules/blog/post-snippet/post-snippet.module.scss
