+++ modules/blog/posts-list/index.tsx
@@ -1,0 +1,3 @@
+import PostsList from './posts-list';
+
+export default PostsList;
