+++ modules/blog/posts-list/posts-list.module.scss
