+++ utils/date-formatting.ts
@@ -1,5 +1,5 @@
 export const dateFormatting = (date: string) => {
-  return new Date(date).toLocaleDateString('default', {
+  return new Date(date).toLocaleDateString('en-GB', {
     day: 'numeric',
     month: 'short',
     year: 'numeric',
